Unified HMI: Add wayland-ivi-extension support
The distributed display framework needs Wayland IVI Extension to
control the layout of application. So, it is necessary to add
wayland-ivi-extension until the support to gRPC in agl-compositor
is completed.

Additionally, add uhmi-ivi-wm to apply ivi-layer and ivi-surface layout
to screen using ilm API of ivi-extension based on ULA protocol.

known issue:
  There is an issue where touch input is not working with the
  wayland-ivi-extension, but this is unrelated to UHMI and occurs with
  the wayland-ivi-extension alone in libweston v13.

Bug-AGL: SPEC-5254

Change-Id: I08c917b36904ed6e9d37369d1f6343bd2454013f
Signed-off-by: Kenta <murakami.kenta002@jp.panasonic.com>